+++ android/app/src/main/java/com/example/todo/model/ActionType.kt
@@ -1,10 +1,6 @@
-package com.example.todo.common
+package com.example.todo.model
 
 import com.example.todo.R
-
-enum class ProgressType(val value: Int) {
-    TO_DO(R.string.progress_todo), IN_PROGRESS(R.string.progress_ing), DONE(R.string.progress_done)
-}
 
 enum class ActionType(val value: Int) {
     ADD(R.string.action_add), REMOVE((R.string.action_remove)), MOVE((R.string.action_move)), UPDATE(
