+++ android/app/src/main/java/com/example/todo/model/ProgressType.kt
@@ -1,0 +1,7 @@
+package com.example.todo.model
+
+import com.example.todo.R
+
+enum class ProgressType(val value: Int) {
+    TO_DO(R.string.progress_todo), IN_PROGRESS(R.string.progress_ing), DONE(R.string.progress_done)
+}
