+++ android/app/src/main/java/com/example/todo/respository/ActionLogDataSource.kt
@@ -1,6 +1,5 @@
 package com.example.todo.respository
 
-import com.example.todo.model.ActionLog
 import com.example.todo.network.JsonActionLog
 
 interface ActionLogDataSource {
