+++ android/app/src/main/java/com/example/todo/respository/ActionLogRemoteDataSource.kt
@@ -1,11 +1,5 @@
 package com.example.todo.respository
 
-import android.content.Context
-import android.util.Log
-import com.example.todo.R
-import com.example.todo.common.ActionType
-import com.example.todo.common.ProgressType
-import com.example.todo.model.ActionLog
 import com.example.todo.network.JsonActionLog
 import com.example.todo.network.RetrofitClient
