+++ android/app/src/main/java/com/example/todo/respository/ToDoRepository.kt
@@ -1,6 +1,4 @@
 package com.example.todo.respository
 
-import android.content.Context
-
 class ToDoRepository(private val toDoDataSource: ToDoDataSource) {
 }
